requirements.txt
faker
fastapi
pydantic
uvicorn[standard]
databases
sqlalchemy
asyncpg
psycopg2-binary